still more debugging output
git-svn-id: svn://localhost/ardour2/branches/3.0@13580 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/evoral/src/Curve.cpp
@@ -257,7 +257,7 @@ Curve::_get_vector (double x0, double x1, float *vec, int32_t veclen)
 		double frac = (x1 - max_x) / (x1 - x0);
 		
 		cerr << "compute subveclen from " << original_veclen << " * " << frac
-		     << " taken from " << x0 << " .. " << x0 
+		     << " taken from " << (int64_t) x0 << " .. " << (int64_t) x0 
 		     << endl;
 
 		int64_t subveclen = (int64_t) floor (original_veclen * frac);
